Function having Default Arguement Values
The fundamental difference between a Parameter and the argument is that the Parameter is the variable name assigned to the function especially while defining the function. Eg; func(number): number**3. The argument is the value that is assigned to that function. Eg; while calling the func(3). Here, 3 is the argument passed during the function call and the value is assigned to the Parameter 'number'.