Anonymous or Lambda functions
This code contain a simple example of lambda function. The lambda is keyword that represents the function without a name. In order, to use the lambda. You need to pass the argument to the object which deals with lambda function. eg; double= lambda x: x*2 , we need to call as double(15)