list as Queue and Stack
The code contains the use case of list. list as Queue and list as Stack. Stack can be imagined as the Vertical list with the concept of "First in,Last out" or can say "Last in,First Out" when calling pop() method. Queue can be imagined as a horizontal list with the concept of "first in, first out". The method use for queue is to import the module named collections and method used is deque(). And to pop is popleft()